Use three dashes instead of explicit em-dash
Sphinx can automatically convert this to the correct HTML entity
using SmartyPants. However, there is a bug with the Guzzle theme
preventing this configuration option to work. I fixed this bug on
my fork of the Guzzle theme, so this PR also switches to using this
branch until my branch is merged and released in the Guzzle theme.