Add governance document
This summarizes the policies that we've been following to this
point, as well as some recent changes that have been discussed
at development meetings.

One new concept in this document are the three roles that team
members can have. All team members are contributors; people with
the inclination and time are reviewers; and project leads
are maintainers.