Fix links to underscored examples
We use hyphens in example filenames now. Also, the notebook in the
getting started section should be directly downloaded if possible.
Using the raw.githubusercontent.com link links to the raw notebook
source, but Github doesn't set the content types to automatically
download, so in most cases it will just open in a new tab. I added
text to instruct the reader to press Ctrl+s to save the notebook.