Create ecosystem diagram with text
The version of the diagram in `ecosystem.svg` had all text converted
to paths; this version of the diagram continues to have text as
paths, and is the version used on the website.

The version in `ecosystem-text.svg` has all text as text (requires
the "IBM Plex Sans" font to render correctly). It should be used to
generate `ecosystem.svg` by converting text to paths.